fix(online-eval): address session queue review
